+++ packages/easy-antd-modal-mobile/.fatherrc.ts
