+++ vitest.config.ts
@@ -1,3 +1,4 @@
+import { resolve } from 'path';
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
@@ -9,6 +10,9 @@ export default defineConfig({
     css: true,
     alias: {
       '##': __dirname,
+      'easy-antd-modal': resolve(__dirname, 'packages/easy-antd-modal/src'),
+      'easy-antd-modal-m': resolve(__dirname, 'packages/easy-antd-modal-mobile/src'),
+      '@wuxh/use-modal-enhanced': resolve(__dirname, 'packages/hooks/src'),
     },
     coverage: {
       reporter: ['text', 'text-summary', 'json', 'lcov'],
